:judge: :copyright: legal: Clarify copyright for DPG certification
This commit adds a new note to the README file about copyright ownership
as documented by the GPLv3 in its "How to Apply These Terms to Your New
Programs" section:

https://github.com/cboard-org/cboard/blob/e899dcac56b456794f5f9407d202513fd68b9433/LICENSE#L623

Of note, it clarifies that the intellectual property is stewarded by the
OTTAA Project, following the recent partnership of the two projects:

* https://github.com/unicef/publicgoods-candidates/pull/559#pullrequestreview-815340890
* https://twitter.com/UNICEFinnovate/status/1429787336368431104
* https://unicef-if.discourse.group/t/meet-ottaa-project-tech-for-speech-impairment/132?u=jwf

This clarifies feedback that came up in review of Cboard's application
to become a Digital Public Good, and will enable the certification
process to move forward. :rocket:

See discussion in unicef/publicgoods-candidates#559 for more context.

CC: @martinbedouret @hectoritr @nathanbaleeta

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

:beetle: licenses: Note Mulberry Symbol set licenses

Actually uses CC BY-SA 4.0 and not CC BY-SA 2.0 UK, see license:

https://github.com/mulberrysymbols/mulberry-symbols/blob/master/LICENSE.txt

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>